refactor: concertProps > concertType 타입명칭 변경
